% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fkcentroids-package.R
\docType{package}
\name{fkcentroids-package}
\alias{fkcentroids-package}
\alias{fkcentroids}
\title{fkcentroids: Functional K-Centroids Clustering Using Phase and Amplitude Components}
\description{
Cluster functional data using phase and amplitude components of each
function. By weighting phase and amplitude variation differently, clustering
results can be obtained from multiple perspectives. Routines for
synchronization, \if{html}{\out{<i>}}k\if{html}{\out{</i>}}-means clustering, and \if{html}{\out{<i>}}k\if{html}{\out{</i>}}-medians clustering are provided.
Details are in Kang and Oh (2026).
}
\references{
Kang S. and Oh H.-S. (2026) \dQuote{Multiview functional
clustering using latent representations of phase and amplitude components,}
\emph{Unpublished Manuscript}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seungwoo-stat/fkcentroids}
  \item Report bugs at \url{https://github.com/seungwoo-stat/fkcentroids/issues}
}

}
\author{
\strong{Maintainer}: Seungwoo Kang \email{kangsw@skku.edu} (\href{https://orcid.org/0000-0001-8082-0794}{ORCID})

Authors:
\itemize{
  \item Hee-Seok Oh \email{heeseok@stats.snu.ac.kr}
}

}
\keyword{internal}
