% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fkmedians.R
\name{fkmedians}
\alias{fkmedians}
\alias{fkmedians_pre}
\alias{print.fkmedians}
\alias{fitted.fkmedians}
\title{title}
\usage{
fkmedians_pre(X.clrv, Y, t, alpha.scale = 1, k, niter = 20, nstart = 1)

fkmedians()
}
\description{
title

title
}
