% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{syncftn}
\alias{syncftn}
\alias{auc_sync}
\alias{fr_sync}
\alias{plot.syncftn}
\title{Time-Synchronizing Mappings}
\usage{
auc_sync(Ytilde, x, t, p = 1)

fr_sync(Ytilde, x, t, template)

\method{plot}{syncftn}(x, phase_mode = c("raw", "clrv"), ...)
}
\arguments{
\item{Ytilde}{A \if{html}{\out{<i>}}m\if{html}{\out{</i>}}\eqn{\times}\if{html}{\out{<i>}}n\if{html}{\out{</i>}} matrix whose \if{html}{\out{<i>}}i\if{html}{\out{</i>}}th column
contains the values of the \if{html}{\out{<i>}}i\if{html}{\out{</i>}}th observed function evaluated at the
\if{html}{\out{<i>}}m\if{html}{\out{</i>}} time points \code{x}.}

\item{x}{A \code{syncftn} object, obtained as a result of the functions
\code{auc_sync()} or \code{fr_sync()}.}

\item{t}{A numeric vector of length \if{html}{\out{<i>}}T\if{html}{\out{</i>}} giving the time points at which
the phase and amplitude components are evaluated. This vector must start at
0 and end at 1.}

\item{p}{A numeric value specifying the power parameter of the AUC
time-synchronizing mapping. The default is \code{p = 1}.}

\item{template}{A numeric vector of length \if{html}{\out{<i>}}T\if{html}{\out{</i>}} giving the template
function value evaluated at the time points \code{t}.}

\item{phase_mode}{A character string.
\itemize{
\item \code{raw} (the default): Plots phase components in their original form.
\item \code{clrv}: Plots phase components after centered log-ratio velocity transformation. Refer to \code{\link[=X2Xclrv]{X2Xclrv()}}.
}}

\item{...}{Further graphical parameters supplied to the internal
\code{\link[graphics:matplot]{graphics::matplot()}} function. \code{main} and \code{ylab} arguments are ignored.}
}
\value{
\code{auc_sync()} and \code{fr_sync()} returns a object of class
\code{syncftn}, which is a list containing the following components,
obtained from AUC synchronization and FR synchronization, respectively:
\itemize{
\item{\code{X}}: A \if{html}{\out{<i>}}T\if{html}{\out{</i>}}\eqn{\times}\if{html}{\out{<i>}}n\if{html}{\out{</i>}} matrix of phase components evaluated
at the time points \code{t}.
\item{\code{Y}}: A \if{html}{\out{<i>}}T\if{html}{\out{</i>}}\eqn{\times}\if{html}{\out{<i>}}n\if{html}{\out{</i>}} matrix of amplitude components
evaluated at the time points \code{t}.
}

\code{plot.syncftn()} plots phase and amplitude components of each
observed function.
}
\description{
Area-under-the-curve (AUC) time-synchronizing mapping and Fisher--Rao (FR)
time-synchronizing mapping.
}
\details{
Let \eqn{\tilde{Y}(x)} be an observed function defined on \eqn{x \in [T_1,
T_2]}. The AUC time-synchronizing mapping is defined as
\deqn{
\varphi_{\tilde{Y}}(x)
:= \left( \frac{\int_{T_1}^x |\tilde{Y}(s)|^p \, \mathrm{d}s}
     {\int_{T_1}^{T_2} |\tilde{Y}(s)|^p \, \mathrm{d}s} \right)^{1/p},
\quad x \in [T_1, T_2]. }
For further details, see Liu and Müller (2004) and Kang and Oh (2026).

The FR time-synchronizing mapping is defined as
\deqn{
\varphi_{\tilde{Y}} := \operatorname*{argmin}_{\varphi} d(\tilde{Y}\circ \varphi^{-1}, Y_0),}
where
\deqn{
d(Y_1,Y_2):=\left\|\operatorname{sgn}(DY_1)\sqrt{\lvert DY_1\rvert} - \operatorname{sgn}(DY_2)\sqrt{\lvert DY_2\rvert} \right\|_2,}
with \eqn{\operatorname{sgn}(u)=1} if \eqn{u\ge 0} and \eqn{-1} otherwise.
The phase component is then defined as \eqn{X(t) :=
\varphi_{\tilde{Y}}^{-1}(t)}, and the amplitude
component is defined as \eqn{Y(t) := (\tilde{Y} \circ
\varphi_{\tilde{Y}}^{-1})(t)} for \eqn{t\in[0,1]}. For further details, see
Srivastava et al. (2011) and Kang and Oh (2026).
}
\examples{
t <- seq(0, 1, length.out = 100)
sync <- auc_sync(seoul_bike$Ytilde[,1:10], seoul_bike$x, t)
plot(sync)
par(mfrow = c(1,2))
plot(sync, col = 1)

template <- 5 * dnorm(t, 0.2, 0.1) + 5 * dnorm(t, 0.8, 0.1)
sync <- fr_sync(seoul_bike$Ytilde[,1:10], seoul_bike$x, t, template)
plot(sync, col = 1)
lines(t, template, col = 2)

}
\references{
Kang S. and Oh H.-S. (2026) \dQuote{Multiview representation and clustering of
functional data,} \if{html}{\out{<i>}}Unpublished Manuscript\if{html}{\out{</i>}}.

Liu X. and Müller H.-G. (2004).\dQuote{"Functional convex averaging and
synchronization for time-warped random curve,} \if{html}{\out{<i>}}Journal of the American
Statistical Association\if{html}{\out{</i>}}, \if{html}{\out{<b>}}99\if{html}{\out{</b>}}(467), 687--699.

Srivastava A., Wu W., Kurtek S., Klassen E., and Marron J. S. (2011)
\dQuote{Registration of functional data using Fisher--Rao metric,} \if{html}{\out{<i>}}arXiv preprint
arXiv:1103.3817\if{html}{\out{</i>}}.
}
\seealso{
\code{\link[fdasrvf:pair_align_functions]{fdasrvf::pair_align_functions()}} for FR synchronization method.
\code{\link[=X2Xclrv]{X2Xclrv()}} for centered log-ratio velocity transformation.  \code{\link[=fkmeans]{fkmeans()}}
and \code{\link[=fkmedians]{fkmedians()}} for \eqn{k}-centroids clustering using phase and
amplitude components.
}
