% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aitchison.R
\name{X2Xclrv}
\alias{X2Xclrv}
\alias{X2Xclrv.matrix}
\alias{X2Xclrv.syncftn}
\alias{plot.Xclrv}
\title{Centered Log-Ratio Velocity Transformation}
\usage{
X2Xclrv(X, t)

\method{X2Xclrv}{matrix}(X, t)

\method{X2Xclrv}{syncftn}(X, t = attr(X, "t"))

\method{plot}{Xclrv}(x, ...)
}
\arguments{
\item{X}{A matrix representing phase components, or an object of class
\code{syncftn}. If a matrix is provided, it must be a
\ifelse{html}{\out{<i>T</i>}\eqn{\times}\out{<i>n</i>}}{\eqn{T\times n}} matrix of \if{html}{\out{<i>}}n\if{html}{\out{</i>}} phase components evaluated
at the time points \code{t}.}

\item{t}{A numeric vector of length \ifelse{html}{\out{<i>T</i>}}{\eqn{T}} giving the time points at which
the phase components are evaluated. This vector must start at 0 and end at
1.}

\item{x}{A \code{Xclrv} object, obtained as a result of the function
\code{X2Xclrv()}.}

\item{...}{Further graphical parameters supplied to the internal
\code{\link[graphics:matplot]{graphics::matplot()}} function.}
}
\value{
\code{X2Xclrv()} returns an object of class \code{Xclrv}, which is a
\ifelse{html}{(\out{<i>T</i>}--1)\eqn{\times}\out{<i>n</i>}}{\eqn{(T-1)\times n}} matrix of clrv transformed phase components
evaluated over the intervals defined by the time points \code{t}.

\code{plot.Xclrv()} plots phase components after centered log-ratio
velocity transformation.
}
\description{
Transform phase components using centered log-ratio velocity (clrv)
transformation.
}
\details{
Let \eqn{X(t)} be a phase component defined on \eqn{t \in [0,1]}. centered log-ratio velocity (clrv)
transformation is defined as
\deqn{
\texttt{clrv}(X)(t) := \log(DX(t)) - \int_0^1 \log(DX(s))\mathrm{d}s,
\quad t \in [0, 1], }
where \eqn{D} is a differential operator.
}
\examples{
t <- seq(0, 1, length.out = 100)
sync <- auc_sync(seoul_bike$Ytilde[,1:10], seoul_bike$x, t)
plot(X2Xclrv(sync))

}
\references{
Kang S. and Oh H.-S. (2026) \dQuote{Multiview functional
clustering using latent representations of phase and amplitude components,}
\emph{Unpublished Manuscript}.
}
\seealso{
\code{\link[=auc_sync]{auc_sync()}} and \code{\link[=fr_sync]{fr_sync()}} for time-synchronizing mappings.
}
