% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aitchison.R
\name{Xclrv2X}
\alias{Xclrv2X}
\title{Inverse Centered Log-Ratio Velocity Transformation}
\usage{
Xclrv2X(Xclrv, t, x)
}
\arguments{
\item{Xclrv}{A \ifelse{html}{(\out{<i>T</i>}--1)\eqn{\times}\out{<i>n</i>}}{\eqn{(T-1)\times n}} matrix of clrv transformed phase components
evaluated over the intervals defined by the time points \code{t}.}

\item{t}{A numeric vector of length \ifelse{html}{\out{<i>T</i>}}{\eqn{T}} giving the time points at which
the phase components are evaluated. This vector must start at 0 and end at
1.}

\item{x}{A numeric vector of length \ifelse{html}{\out{<i>m</i>}}{\eqn{m}}
giving the observed time points.}
}
\value{
A \ifelse{html}{\out{<i>T</i>}\eqn{\times}\out{<i>n</i>}}{\eqn{T\times n}} matrix of inverse-clrv transformed phase components
evaluated at the time points \code{t}.
}
\description{
Inverse of the centered log-ratio velocity (clrv) transformation.
}
\details{
Let \eqn{X(t)} be a phase component defined on \eqn{t \in [0,1]}. centered log-ratio velocity (clrv)
transformation is defined as
\deqn{
\texttt{clrv}(X)(t) := \log(DX(t)) - \int_0^1 \log(DX(s))\mathrm{d}s,
\quad t \in [0, 1], }
where \eqn{D} is a differential operator.

If \eqn{X: [0,1]\to[T_1,T_2]} for \eqn{T_1 < T_2}, the inverse of the clrv transformation is defined as
\deqn{
T_1 + (T_2 - T_1)\frac{\int_0^t \exp\{\texttt{clrv}(X)(s)\}\mathrm{d}s}{\int_0^1 \exp\{\texttt{clrv}(X)(s)\}\mathrm{d}s},\quad t\in[0,1].
}
}
\examples{
t <- seq(0, 1, length.out = 100)
sync <- auc_sync(seoul_bike$Ytilde[,1:10], seoul_bike$x, t)
xclrv <- X2Xclrv(sync)
range(Xclrv2X(xclrv, t, seoul_bike$x) - sync$X)

}
\references{
Kang S. and Oh H.-S. (2026) \dQuote{Multiview representation and clustering of
functional data,} \emph{Unpublished Manuscript}.
}
\seealso{
\code{\link[=X2Xclrv]{X2Xclrv()}} for clrv transformation.
}
