% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fkmeans.R
\name{fkmeans}
\alias{fkmeans}
\alias{fkmeans_pre}
\alias{print.fkmeans}
\alias{fitted.fkmeans}
\title{Functional \eqn{k}-Means Clustering Using Phase and Amplitude
Components}
\usage{
fkmeans_pre(
  Xclrv,
  Y,
  t,
  alpha_scale = 1,
  k,
  itermax = 10,
  nstart = 1,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  trace = FALSE
)

fkmeans(
  Ytilde,
  x,
  t,
  sync_map = c("auc", "fr"),
  sync_args,
  alpha_scale,
  k,
  itermax = 10,
  nstart = 1,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  trace = FALSE
)

\method{fitted}{fkmeans}(object, method = c("centers", "classes"), ...)
}
\arguments{
\item{Xclrv}{A (\if{html}{\out{<i>}}T\if{html}{\out{</i>}}--1)\eqn{\times}\if{html}{\out{<i>}}n\if{html}{\out{</i>}} matrix of centered log-ratio
velocity transformed phase components evaluated over the intervals defined
by the time points \code{t}. Refer to \code{\link[=X2Xclrv]{X2Xclrv()}}.}

\item{Y}{A \if{html}{\out{<i>}}T\if{html}{\out{</i>}}\eqn{\times}\if{html}{\out{<i>}}n\if{html}{\out{</i>}} matrix of amplitude components
evaluated at the time points \code{t}.}

\item{t}{A numeric vector of length \if{html}{\out{<i>}}T\if{html}{\out{</i>}} giving the time points at which
the phase and amplitude components are evaluated. This vector must start at
0 and end at 1.}

\item{alpha_scale}{A numeric indicating the value of multiview parameter. The
multiview parameter \eqn{\alpha} is set to \eqn{\alpha = \alpha_0 \times
  }\code{alpha_scale}. See the details below. By default, set to 1.}

\item{k}{A numeric indicating the number of clusters.}

\item{itermax}{A numeric indicating the maximum number of iterations allowed
in the \eqn{k}-means algorithm. By default, set to 10.}

\item{nstart}{A numeric indicating the number of initial sets. By default, set to 1.}

\item{algorithm}{A character string indicating the algorithm for \eqn{k}-means
clustering. \code{"Hartigan-Wong"} algorithm is set as default.}

\item{trace}{A boolean. If \code{TRUE} and \code{algorithm ==
"Hartigan-Wong"}, it prints tracing information on the console.}

\item{Ytilde}{A \if{html}{\out{<i>}}m\if{html}{\out{</i>}}\eqn{\times}\if{html}{\out{<i>}}n\if{html}{\out{</i>}} matrix whose \if{html}{\out{<i>}}i\if{html}{\out{</i>}}th column
contains the values of the \if{html}{\out{<i>}}i\if{html}{\out{</i>}}th observed function evaluated at the
\if{html}{\out{<i>}}m\if{html}{\out{</i>}} time points \code{x}.}

\item{x}{A numeric vector of length \if{html}{\out{<i>}}m\if{html}{\out{</i>}} giving the observed time points
corresponding to \code{Ytilde}.}

\item{sync_map}{A character string. If \code{"auc"}, AUC time-synchronizing
mapping is used. If \code{"fr"}, FR time-synchronizing mapping is used.
Refer to \code{\link[=auc_sync]{auc_sync()}} and \code{\link[=fr_sync]{fr_sync()}}.}

\item{sync_args}{If \code{sync_map == "auc"} it represents a numeric
indicating the parameter \eqn{p} used in AUC time-synchronizing mapping. If
\code{sync_map == "fr"} it represent the template function used in FR
time-synchronizing mapping.}

\item{...}{Not used.}

\item{object}{A \code{fkmeans} object, obtained as a result of the function
\code{fkmeans()}.}

\item{method}{A character string.
\itemize{
\item centers: Returns cluster centers for each curve.
\item classes: Returns a vector of class assignments.
}}
}
\value{
\code{fkmeans_pre()} and \code{fkmeans()} return an object of class
\code{fkmeans}, which is a list containing the following components:
\itemize{
\item{\code{cluster}}: A vector of integers (from \code{1:k}) indicating the cluster to which each function is allocated.
\item{\code{centers.Xclrv}}: A (\if{html}{\out{<i>}}T\if{html}{\out{</i>}}--1)\eqn{\times}\if{html}{\out{<i>}}k\if{html}{\out{</i>}} matrix of phase components' cluster centers (centered log-ratio velocity transformed).
\item{\code{centers.Y}}: A \if{html}{\out{<i>}}T\if{html}{\out{</i>}}\eqn{\times}\if{html}{\out{<i>}}k\if{html}{\out{</i>}} matrix of amplitude components' cluster centers.
\item{\code{totss}}: The total sum of squares.
\item{\code{withinss}}: A vector of within-cluster sum of squares, one component per cluster.
\item{\code{tot.withinss}}: Total within-cluster sum of squares, i.e., \code{sum(withinss)}.
\item{\code{betweenss}}: The between-cluster sum of squares, i.e. \code{totss-tot.withinss}.
\item{\code{size}}: The number of functions in each cluster.
\item{\code{iter}}: The number of (outer) iterations.
\item{\code{ifault}}: Indicator of a possible algorithm problem; refer to \code{\link[stats:kmeans]{stats::kmeans()}}.
\item{\code{alpha0}}: The reference value \eqn{\alpha_0}.
}

\code{print()} and \code{fitted()} methods are supported for the
object of class \code{fkmeans}. \code{fitted.fkmeans()} with \code{method =
  "centers"} returns cluster centers (one for each input point) and
\code{method = "classes"} returns a vector of class assignments.
}
\description{
Conducts \eqn{k}-means clustering by jointly considering phase and amplitude
variation. The relative importance of the two components can be explicitly
controlled by the user via the multiview parameter \eqn{\alpha}. See the
details below.
}
\details{
The distance between two observed functions is defined in terms of their
phase and amplitude components. For two functions with components
\eqn{(X_1, Y_1)} and \eqn{(X_2, Y_2)}, the distance is given by
\deqn{
\left\{
  \alpha \|\texttt{clrv}(X_1) - \texttt{clrv}(X_2)\|_2^2
  + \|Y_1 - Y_2\|_2^2
\right\}^{1/2},
}
where \eqn{\|\cdot\|_2} denotes the usual \eqn{\mathbb{L}^2} norm and
\eqn{\alpha \ge 0} is the multiview parameter. Based on this distance,
\eqn{k}-means clustering is performed.

A reference value \eqn{\alpha_0}, which serves as a baseline around which
\eqn{\alpha} may be varied, is selected as follows. The value \eqn{\alpha_0}
is defined as the ratio of the total sum of squares of the amplitude
components to that of the phase components. See the documentation for
\code{\link[stats:kmeans]{stats::kmeans()}} and Kang and Oh (2026) for further details.
}
\examples{
t <- seq(0, 1, length.out = 100)
sync <- auc_sync(seoul_bike$Ytilde[,1:10], seoul_bike$x, t)
fkmeans_pre(X2Xclrv(sync), sync$Y, t, alpha_scale = 1, k = 2)
fkmeans(seoul_bike$Ytilde[,1:10], seoul_bike$x, t, sync_map = "auc",
  sync_args = 1, alpha_scale = 1, k = 2)

}
\references{
Kang S. and Oh H.-S. (2026) \dQuote{Multiview representation and clustering of
functional data,} \if{html}{\out{<i>}}Unpublished Manuscript\if{html}{\out{</i>}}.
}
\seealso{
\code{\link[stats:kmeans]{stats::kmeans()}} for multivariate \eqn{k}-means clustering.
\code{\link[=fkmedians_pre]{fkmedians_pre()}} and \code{\link[=fkmedians]{fkmedians()}} for robust functional \eqn{k}-medians
clustering. \code{\link[=auc_sync]{auc_sync()}} and \code{\link[=fr_sync]{fr_sync()}} for time-synchronizing mappings.
\code{\link[=X2Xclrv]{X2Xclrv()}} for centered log-ratio velocity transformation.
}
